Sections 5 & 7-VertexAI-Development/Bikeshare-Model/Cloud-Run-Flask-Online-Prediction/requirements.txt
Flask>=3.0.0
gunicorn>=22.0.0
google-cloud-aiplatform>=1.60.0
